requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mixin==1.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
